requirements.txt
blinker==1.9.0
click==8.1.8
colorama==0.4.6
Flask==3.1.1
flask-cors==6.0.1
gunicorn==23.0.0
importlib_metadata==8.7.0
itsdangerous==2.2.0
Jinja2==3.1.6
MarkupSafe==3.0.2
packaging==25.0
psycopg2-binary==2.9.10
Werkzeug==3.1.3
zipp==3.23.0
